+++ apps/web/modules/event-types/views/event-types-single-view.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { EventType } from "@calcom/features/eventtypes/components/event-type/EventType";
+import { EventType } from "@calcom/features/eventtypes/components/EventType";
 import type { EventTypeSetupProps } from "@calcom/features/eventtypes/lib/types";
 
 /* eslint-disable @typescript-eslint/no-empty-function */
